Better Ruby Haml compatibility for attributes.
Currently the attribute generation doesn't allow empty attributes
and removes them, but there might be legitim cases for using empty
attributes.

    %input{ name: '' }

was previously rendered to

    <input>

but from now this will produce

    <input name=''>

Also the boolean attribute logic was too much decoupled from the tag
generation, so there was no distinction between regular attributes
and boolean attributes at tag generation time.

    %input{ name: 'name' }

was rendered to

    <input name>

but since it's not a boolean attribute, it must be rendered to

    <input name='name'>