Extract render function from the template generation.
This seperates the function that generates the HTML from to JST and
enables the introduction of a simple HamlCoffee facade for getting the
actual JST template or just the render function without any JST related
stuff.

The new facade enables Haml Coffee to be used as Expresso view and other
tools can handle JST generation on their own, for example
haml_coffee_assets could just pass function to the Sprockets JST
processor.