Rewrite attribute parsing to be more robust. (Closes #26)
This rewrites the HAML attribute tag parsing to move some parse logic
from the regexp to the code, which allows more robust parsing when it
comes to inline CoffeeScript detection.

Previously the attributes key and value are purely parsed with a regexp,
expect the data attributes.

The rewrite combines the data attribute parsing and the regexp based
attribute parsing by simply splitting the attributes by the chosen key
format and further process the key and values.

This allows more complex CoffeeScript in the attributes like

    %a{ href: [1, 2].join() } OMG

which was not possible before.